Create Batch Effect Correction.R
To ensure that datasets from the various cohorts are compatible for cross-dataset analysis and integration, a t-Distributed Stochastic Neighbour Embedding (t-SNE) of the ASVs was used to reveal biases in the data samples cluster within the original studies. To overcome this challenge, the Harman function in the Harman package in R was used to correct batch effects that may be introduced. Harman is a PCA and constrained optimization-based technique that maximizes the removal of batch effects from datasets. The constraint is that the probability of overcorrection (i.e., removing genuine biological signal and batch noise) is kept to a fraction set by the end-user. pcaPlot was employed to visualize the removed batch effects.